Produce better errors for unsupported plugins
Previously, if an unsupported plugin was used on the command line along
with its options (e.g. tmt run provision -h virtual -i image), the
plugin option would be assigned by click to provision instead resulting
in a confusing message (Unknown option for provision). This was an issue
with provision and report since they have installable plugins.

Allow the steps to consume unknown options and process them later
manually (instead of leaving the work for click) based on how. This
allows us to give a hint about installing the correct plugins.

Signed-off-by: František Nečas <fnecas@redhat.com>